backend/requirements.txt
-i https://pypi.org/simple
amqp==5.3.1; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
billiard==4.2.1; python_version >= '3.7'
celery==5.4.0; python_version >= '3.8'
click==8.1.8; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
django==5.1.6; python_version >= '3.10'
django-cors-headers==4.7.0; python_version >= '3.9'
django-environ==0.12.0; python_version >= '3.9' and python_version < '4'
flower==2.0.1; python_version >= '3.7'
humanize==4.12.1; python_version >= '3.9'
kombu==5.4.2; python_version >= '3.8'
mysqlclient==2.2.7; python_version >= '3.8'
prometheus-client==0.21.1; python_version >= '3.8'
prompt-toolkit==3.0.50; python_full_version >= '3.8.0'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2025.1
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.3; python_version >= '3.8'
tornado==6.4.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
